+++ src/error_chain.rs
@@ -251,7 +251,7 @@ macro_rules! impl_error_chain_processed {
         // The ErrorKind type
         // --------------
 
-        quick_error! {
+        impl_error_chain_kind! {
             /// The kind of an error.
             #[derive(Debug)]
             pub enum $error_kind_name {